+++ src/app/admin/components/BottomSheetWrapper.tsx
@@ -1,0 +1,50 @@
+"use client";
+
+import { useEffect, useRef, ReactNode } from "react";
+
+import { motion, AnimatePresence } from "framer-motion";
+
+interface BottomSheetWrapperProps {
+  onClose: () => void;
+  children: ReactNode;
+}
+
+export default function BottomSheetWrapper({
+  onClose,
+  children,
+}: BottomSheetWrapperProps) {
+  const modalRef = useRef<HTMLDivElement>(null);
+
+  useEffect(() => {
+    const handleClickOutside = (e: MouseEvent) => {
+      if (modalRef.current && !modalRef.current.contains(e.target as Node)) {
+        onClose();
+      }
+    };
+    document.addEventListener("mousedown", handleClickOutside);
+    return () => document.removeEventListener("mousedown", handleClickOutside);
+  }, [onClose]);
+
+  return (
+    <AnimatePresence>
+      <motion.div
+        className="fixed inset-0 bg-black/30 z-50 flex items-end justify-center"
+        initial={{ opacity: 0 }}
+        animate={{ opacity: 1 }}
+        exit={{ opacity: 0 }}
+        transition={{ duration: 0.15 }}
+      >
+        <motion.div
+          ref={modalRef}
+          className="bg-white w-full max-w-md rounded-t-[40px] p-5 shadow-lg"
+          initial={{ y: 50 }}
+          animate={{ y: 0 }}
+          exit={{ y: 50 }}
+          transition={{ duration: 0.2, ease: "easeOut" }}
+        >
+          {children}
+        </motion.div>
+      </motion.div>
+    </AnimatePresence>
+  );
+}
